@@ -7,36 +7,44 @@ using MongoFramework.Infrastructure.Diagnostics;
 using MongoFramework.Utilities;
 
 // ReSharper disable once CheckNamespace
-namespace MongoFramework
+namespace MongoFramework;
+
+/// <summary>
+/// A MongoDbConnection that accepts a TenantInfo and uses its connection string to create the Data Context
+/// </summary>
+public interface IMongoPerTenantConnection: IMongoDbConnection
 {
-    /// <summary>
-    /// A MongoDbConnection that accepts a TenantInfo and uses its connection string to create the Data Context
-    /// </summary>
-    public interface IMongoPerTenantConnection: IMongoDbConnection
+    ITenantInfo TenantInfo { get; }
+}
+
+public class MongoPerTenantConnection : MongoDbConnection, IMongoPerTenantConnection
+{
+    public ITenantInfo TenantInfo { get; }
+
+    public MongoPerTenantConnection(ITenantInfo ti, IOptions<MongoPerTenantConnectionOptions> options = null)
     {
-        ITenantInfo TenantInfo { get; }
+        Check.NotNull(ti, nameof(ti));
+        TenantInfo = ti;
+        if (IsMongoDbConnectionString(ti.ConnectionString))
+        {
+            Url = new MongoUrl(ti.ConnectionString);
+        }
+        else if (IsMongoDbConnectionString(options?.Value?.DefaultConnectionString))
+        {
+            Url = new MongoUrl(options.Value.DefaultConnectionString);
+        }
+        else
+        {
+            throw new ArgumentException("Connection String required.");
+        }
     }
 
-    public class MongoPerTenantConnection : MongoDbConnection, IMongoPerTenantConnection
+    private static bool IsMongoDbConnectionString(string value)
     {
-        public ITenantInfo TenantInfo { get; }
-
-        public MongoPerTenantConnection(ITenantInfo ti, IOptions<MongoPerTenantConnectionOptions> options = null)
+        if (string.IsNullOrEmpty(value))
         {
-            Check.NotNull(ti, nameof(ti));
-            TenantInfo = ti;
-            if (!string.IsNullOrEmpty(ti.ConnectionString) && ti.ConnectionString.ToLower().StartsWith("mongodb://"))
-            {
-                Url = new MongoUrl(ti.ConnectionString);
-            }
-            else if (!string.IsNullOrEmpty(options?.Value?.DefaultConnectionString) && options.Value.DefaultConnectionString.ToLower().StartsWith("mongodb://"))
-            {
-                Url = new MongoUrl(options.Value.DefaultConnectionString);
-            }
-            else
-            {
-                throw new ArgumentException("Connection String required.");
-            }
+            return false;
         }
+        return value.ToLowerInvariant().StartsWith("mongodb://") || value.ToLowerInvariant().StartsWith("mongodb+srv://");
     }
 }
